Add artifact sharing endpoints
This adds two new API endpoints: one to generate publish tokens for the
sharing portal, which allows the user servers to publish files to Swift,
and another API endpoint that can generate download URLs. The latter may
not always be exposed as a download URL; probably it can just be a
helper function used before spawning a server.