Add custom logout for federated users
This adds a custom logout handler which, when invoked, checks the user's
login state for a flag indicating they are federated. If they are, it
redirects them to a special Keycloak logout page, which invalidates
their Keycloak IdP session.

Note: differently than for the login handler, for this one we actually
overwrite the original handler path with our own handler instead of the
default, which is possible due to handler installation order. This is
necessary b/c clicking "Logout" in JupyterLab always takes the user to
/logout.