Increase cull limits for servers/kernels
Increase the idle timeout for kernels from 3 hours to 1 day. This should
make it easier to come back to a Notebook after a day  and have its
state still there.

Also increase idle server timeout to 3 days, from 1. This should make it
so you can come back after a weekend or a few days and still have any
custom stuff installed.

Also remove the max age. This is less important to enforce now that we
have federation, because we can keep refreshing the user's session for a
while.