Add session refresh handler
This adds a new handler that will attempt to refresh the user's session
by first logging them out of JupyterHub and then redirecting them to
a page specified in the GET parameter. This should trigger a new login
via OIDC and, if they are still logged in to the upstream IdP, which is
likely, they can proceed straight back to where they were.