Add username hash if username is too long
We can only spawn pods with names no longer than 63 characters. Some of
our users have very long usernames. When spawning, we now check that the
theoretical pod name is less than this length, using the same escape
method used elsewhere in the parent kubespawner.

The escapism library is already a requirement by the kubespawner, and so
this should not be adding an additional library.