Fix volume and pod name length
The pre_spawn_hook is called before the pod object is created in
kubernetes. We use this to overwrite the templates used to generate the
pod name, and the volume name, as these are limited to 63 characters.
Note the claim name in the volume is what is actually used to map data
into the container, so this does not affect user data. In fact, I'm 99%
sure the volume name can be static (no templating) since it is under the pod,
which has a unique name, but I'm assuming there might be some reason
jupyterhub has it include the username.